Refactor: restructure card components
