+++ src/components/Card/index.tsx
@@ -1,1 +1,2 @@
-export { IdeaCard } from "./IdeaCard";
+export { IdeaCard } from "./IdeaCard/IdeaCard";
+export { TodoCard } from "./TodoCard/TodoCard";
